Change SAXParser::textHandler parameter type from int to size_t

+++ cocos/scripting/js-bindings/manual/cocos2d_specifics.cpp
@@ -4715,7 +4715,7 @@ void __JSPlistDelegator::endElement(void *ctx, const char *name) {
     }
 }
 
-void __JSPlistDelegator::textHandler(void *ctx, const char *ch, int len) {
+void __JSPlistDelegator::textHandler(void *ctx, const char *ch, size_t len) {
     CC_UNUSED_PARAM(ctx);
     std::string text((char*)ch, 0, len);
